+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.7.0] - 2019-05-30
+### Added
+- Kitchen and Bathroom rooms
+- Using l as a shortcut for look
+
+### Changed
+- Directions to be a variable
+- Look command to not display room name after
+
 ## [0.6.1] - 2019-05-29
 ### Fixed
 - Typos introduced to game text by v0.6.0
@@ -61,7 +70,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - .gitignore
 - Changelog and Readme files
 
-[Unreleased]: https://github.com/afs2015/PepperRPG/compare/v0.6.1....HEAD
+[Unreleased]: https://github.com/afs2015/PepperRPG/compare/v0.7.0....HEAD
+[0.7.0]: https://github.com/afs2015/PepperRPG/compare/v0.6.1...v0.7.0
 [0.6.1]: https://github.com/afs2015/PepperRPG/compare/v0.6.0...v0.6.1
 [0.6.0]: https://github.com/afs2015/PepperRPG/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/afs2015/PepperRPG/compare/v0.4.0...v0.5.0